Implement Sync Lists

+++ src/de/azapps/mirakel/ListsDataSource.java
@@ -6,12 +6,15 @@ package de.azapps.mirakel;
 import java.util.ArrayList;
 import java.util.List;
 
+import org.apache.http.message.BasicNameValuePair;
 import android.content.ContentValues;
 import android.content.Context;
 import android.database.Cursor;
 import android.database.SQLException;
 import android.database.sqlite.SQLiteDatabase;
 import android.util.Log;
+
+import com.google.gson.Gson;
 
 /**
  * @author weiznich
@@ -22,7 +25,7 @@ public class ListsDataSource {
 	private SQLiteDatabase database;
 	private DatabaseHelper dbHelper;
 	private String[] allColumns = { "_id", "name", "sort_by", "created_at",
-			"updated_at" };
+			"updated_at" ,"sync_state"};
 
 	public ListsDataSource(Context context) {
 		dbHelper = new DatabaseHelper(context);
@@ -44,6 +47,7 @@ public class ListsDataSource {
 		ContentValues values = new ContentValues();
 		values.put("name", name);
 		values.put("sort_by", sort_by);
+		values.put("sync_state", Mirakel.SYNC_STATE_ADD);
 		long insertId = database.insert("lists", null, values);
 		Cursor cursor = database.query("lists", allColumns,
 				"_id = " + insertId, null, null, null, null);
@@ -55,14 +59,22 @@ public class ListsDataSource {
 
 	public void saveList(List_mirakle list) {
 		Log.v(TAG, "saveTask");
+		list.setSync_state(list.getSync_state()==Mirakel.SYNC_STATE_ADD?Mirakel.SYNC_STATE_ADD:Mirakel.SYNC_STATE_NEED_SYNC);		
 		ContentValues values = list.getContentValues();
 		database.update("lists", values, "_id = " + list.getId(), null);
 	}
 
 	public void deleteList(List_mirakle list) {
 		long id = list.getId();
-		database.delete("tasks", "list_id = " + id, null);
-		database.delete("lists", "_id = " + id, null);
+		if(list.getSync_state()==Mirakel.SYNC_STATE_ADD){
+			database.delete("tasks", "list_id = " + id, null);
+			database.delete("lists", "_id = " + id, null);
+		}else{
+			ContentValues values=new ContentValues();
+			values.put("sync_state", Mirakel.SYNC_STATE_DELETE);
+			database.update("tasks", values, "list_id = "+id, null);
+			database.update("lists", values, "_id="+id,null);
+		}
 	}
 
 	public List<List_mirakle> getAllLists() {
@@ -74,7 +86,7 @@ public class ListsDataSource {
 				task_count(Mirakel.LIST_DAILY)));
 		lists.add(new List_mirakle(Mirakel.LIST_WEEKLY, "This Week",
 				task_count(Mirakel.LIST_WEEKLY)));
-		Cursor cursor = database.query("lists", allColumns, null, null, null,
+		Cursor cursor = database.query("lists", allColumns, "not sync_state="+Mirakel.SYNC_STATE_DELETE, null, null,
 				null, null);
 		cursor.moveToFirst();
 		while (!cursor.isAfterLast()) {
@@ -99,7 +111,7 @@ public class ListsDataSource {
 	}
 
 	private int task_count(int list_id) {
-		String count = "Select count(_id) from tasks ";
+		String count = "Select count(_id) from tasks where";
 		switch (list_id) {
 		case Mirakel.LIST_ALL:
 			break;
@@ -108,9 +120,10 @@ public class ListsDataSource {
 		case Mirakel.LIST_WEEKLY:
 			return new TasksDataSource(null).getTasks(Mirakel.LIST_WEEKLY, Mirakel.ORDER_BY_ID).size();
 		default:
-			count += "where list_id=" + list_id;
-
+			count += " list_id=" + list_id+" and";
 		}
+		count+=" not sync_state="+Mirakel.SYNC_STATE_DELETE;
+		open();
 		Cursor c = database.rawQuery(count, null);
 		c.moveToFirst();
 		int task_count = c.getInt(0);
@@ -123,8 +136,107 @@ public class ListsDataSource {
 		int id = cursor.getInt(i++);
 		List_mirakle list = new List_mirakle(id, cursor.getString(i++),
 				cursor.getShort(i++), cursor.getString(i++),
-				cursor.getString(i), task_count(id));
+				cursor.getString(i++), task_count(id),cursor.getInt(i));
 		return list;
 	}
+	
+	public void sync_lists(final String email, final String password, final String url) {		
+		new Network(new DataDownloadCommand() {			
+			@Override
+			public void after_exec(String result) {
+				List_mirakle lists_server[]=new Gson().fromJson(result, List_mirakle[].class);
+				List<List_mirakle>lists_local=getAllLists();
+				for(int i=0; i<lists_local.size();i++){
+					List_mirakle list=lists_local.get(i);
+					switch(list.getSync_state()){
+						case Mirakel.SYNC_STATE_ADD:
+							add_list(list,email,password,url);
+							break;
+						case Mirakel.SYNC_STATE_DELETE:
+							delete_list(list,email,password,url);
+							//TODO Implement Delete from server
+							list.setSync_state(Mirakel.SYNC_STATE_ADD);
+							deleteList(list);
+							list=null;
+							continue;
+						case Mirakel.SYNC_STATE_NEED_SYNC:
+							sync_list(list,email,password,url);
+							break;
+						default:
+					}
+				}
+				merge_with_server(lists_local, lists_server);
+				ContentValues values=new ContentValues();
+				values.put("sync_state", Mirakel.SYNC_STATE_NOTHING);
+				database.update("lists", values, "not sync_state="+Mirakel.SYNC_STATE_NOTHING, null);
+				
+			}
+		},email,password,Mirakel.Http_Mode.GET).execute(url+"/lists.json");	
+	}
 
+	protected void delete_list(final List_mirakle list,final String email,final String password,final String url) {
+		new Network(new DataDownloadCommand() {
+			@Override
+			public void after_exec(String result) {
+				// Do Nothing
+				
+			}
+		},email,password,Mirakel.Http_Mode.DELETE).execute(url+"/lists/"+list.getId()+".json");
+		
+	}
+
+	protected void sync_list(List_mirakle list, String email, String password,
+			String url) {
+		List<BasicNameValuePair> data=new ArrayList<BasicNameValuePair>();
+		data.add(new BasicNameValuePair("list[name]", list.getName()));
+		new Network(new DataDownloadCommand() {
+			@Override
+			public void after_exec(String result) {
+				//Do Nothing
+				
+			}
+		}, email, password, Mirakel.Http_Mode.PUT,data).execute(url+"/lists/"+list.getId()+".json");	
+	}
+
+	protected void add_list(final List_mirakle list,final String email,final String password, final String url) {
+		List<BasicNameValuePair> data=new ArrayList<BasicNameValuePair>();
+		data.add(new BasicNameValuePair("list[name]", list.getName()));
+		new Network(new DataDownloadCommand() {
+			
+			@Override
+			public void after_exec(String result) {
+				List_mirakle list_response=new Gson().fromJson(result, List_mirakle.class);
+				ContentValues values=new ContentValues();
+				values.put("_id", list_response.getId());
+				open();
+				database.update("lists", values, "_id="+list.getId(), null);
+				values=new ContentValues();
+				values.put("list_id", list_response.getId());
+				database.update("tasks", values, "list_id="+list.getId() ,null);
+				
+			}
+		},email,password,Mirakel.Http_Mode.POST,data).execute(url+"/lists.json");
+		
+	}
+
+	protected void merge_with_server(List<List_mirakle> lists,
+			List_mirakle[] lists_server) {
+		for(int i=0;i<lists_server.length;i++){
+			List_mirakle list =getList(lists_server[i].getId());
+			if(list==null){
+				long id=database.insert("lists", null, lists_server[i].getContentValues());
+				ContentValues values=new ContentValues();
+				values.put("_id", lists_server[i].getId());
+				open();
+				database.update("lists", values, "_id="+id, null);
+			}else{
+				if(list.getSync_state()==Mirakel.SYNC_STATE_NOTHING){
+					saveList(lists_server[i]);
+				}else{
+					Log.e(TAG,"Merging lists not implementet");
+				}
+			}
+		}
+		
+	}
 }
